requirements.txt
numpy
pandas
scikit-learn
joblib
regex

fastapi>=0.110,<1.0
uvicorn
python-multipart
pydantic>=2.5,<3.0
annotated-types

requests
jinja2
python-jose[cryptography]

clamd
python-magic
tldextract

psycopg2-binary

